Met en commun le code de disparation de la pièce.
